refactor(test): 拆解 tests/backend 根目录下的大杂烩测试文件
把横跨多个不相关模块的测试按被测源文件归位，路径镜像 apps/backend：

- test_internal_modules.py -> core/memory/markdown/{test_formatting,
  test_consolidation_service}.py、memory2/test_post_response_worker.py
- test_mid_modules.py -> agent/tools/shell/{test_tools,test_validation,
  test_output}.py、agent/tools/test_web_fetch.py、
  memory2/{test_procedure_tagger,test_store}.py
- test_logic_modules.py -> session/test_manager.py、
  proactive_v2/{test_sensor,test_loop}.py
- test_core_decoupling_baseline.py（重构期脚手架）-> agent/core/
  {test_runtime_support,test_passive_support,test_types}.py；
  ToolDiscoveryState 用例从 test_foundation.py 一并归位
- tests/backend/turns/test_orchestrator.py 与 agent/turns/test_orchestrator.py
  重名，合并到后者；重叠的 media 派发用例合并为一个，media 到达
  OutboundDispatch 的断言并入保留用例

一并删除已被既有用例覆盖的重复测试：memory2 的预算耗尽与 explicit
memorized、tool_search 的三个 select/keyword 用例、memory_optimizer 的
optimize/next_tick。tool_search 中未被覆盖的 set_excluded_names 贯通
断言改写后并入 agent/tools/test_search.py。

纯搬运，无行为变更。

Refs #194
