+++ tests/backend/agent/core/test_foundation.py
@@ -3,13 +3,9 @@ from agent.core import (
     ContextBundle,
     InboundMessage,
     LLMResponse,
-    LLMServices,
-    MemoryConfig,
-    MemoryServices,
     OutboundMessage,
     ReasonerResult,
     ToolCall,
-    ToolDiscoveryState,
     TurnRecord,
 )
 
@@ -37,34 +33,3 @@ def test_agent_core_foundation_types_construct_cleanly():
     assert response.tool_calls[0].name == "dummy"
     assert result.invocations[0].id == "c1"
     assert record.reply == "done"
-
-
-def test_agent_core_runtime_support_tool_discovery_lru():
-    state = ToolDiscoveryState(capacity=2)
-    state.update("cli:1", ["tool_a", "tool_b"], {"always"})
-    assert state.get_preloaded("cli:1") == {"tool_a", "tool_b"}
-    assert state.get_preloaded_ordered("cli:1") == ["tool_a", "tool_b"]
-
-    state.update("cli:1", ["tool_a"], {"always"})
-    state.update("cli:1", ["tool_c"], {"always"})
-
-    assert state.get_preloaded("cli:1") == {"tool_a", "tool_c"}
-    assert state.get_preloaded_ordered("cli:1") == ["tool_a", "tool_c"]
-    assert "tool_b" not in state.get_preloaded("cli:1")
-
-
-def test_agent_core_runtime_support_skips_always_on_and_tool_search():
-    state = ToolDiscoveryState()
-    state.update("cli:1", ["always_tool", "tool_search", "hidden_tool"], {"always_tool"})
-
-    assert state.get_preloaded("cli:1") == {"hidden_tool"}
-
-
-def test_agent_core_runtime_support_service_types_hold_objects():
-    llm = LLMServices(provider=object(), light_provider=object())
-    memory = MemoryServices(engine=object())
-    config = MemoryConfig(window=12)
-
-    assert llm.provider is not None
-    assert memory.engine is not None
-    assert config.window == 12
